tags are stings leading to extensible records
