+++ benchmarks/icfp15/pos/DBMovies.Tags.hs
@@ -2,16 +2,14 @@ module MovieClient where
 import DataBase
 
 
-import GHC.CString  -- This import interprets Strings as constants!
-
 import Data.Maybe (catMaybes)
 
 import Prelude hiding (product)
 
 import Control.Applicative ((<$>))
 
-
-type Tag = String 
+data Tag   = Title | Director | Star | Year 
+            deriving (Show, Eq)
 
 data Value = I Int | S Name  
             deriving (Show, Eq)
@@ -33,33 +31,33 @@ data Name = ChickenPlums | TalkToHer | Persepolis | FunnyGames
 
 
 {-@ predicate ValidMovieScheme V = 
-	  ((listElts (ddom V) = Set_cup (Set_sng year) 
-	  	                   (Set_cup (Set_sng star) 
-	  	                   (Set_cup (Set_sng director) 
-	  	                            (Set_sng title))))) @-}
+	  ((listElts (ddom V) = Set_cup (Set_sng Year) 
+	  	                   (Set_cup (Set_sng Star) 
+	  	                   (Set_cup (Set_sng Director) 
+	  	                            (Set_sng Title))))) @-}
 
-{-@ predicate MovieDomain T   = (T = year || T = star || T = director || T = title) @-}
+{-@ predicate MovieDomain T   = (T = Year || T = Star || T = Director || T = Title) @-}
 
-{-@ predicate MovieRange  T V =    (T = year     => ValidYear     V) 
-                                && (T = star     => ValidStar     V) 
-                                && (T = director => ValidDirector V) 
-                                && (T = title    => ValidTitle    V) @-}
+{-@ predicate MovieRange  T V =    (T = Year     => ValidYear     V) 
+                                && (T = Star     => ValidStar     V) 
+                                && (T = Director => ValidDirector V) 
+                                && (T = Title    => ValidTitle    V) @-}
 
-{-@ predicate ValidDirectorScheme V = (listElts (ddom V) = (Set_sng director)) @-} 
-{-@ predicate DirectorDomain T   = ( T = director ) @-}
-{-@ predicate DirectorRange  T V = (T = director => ValidDirector V) @-}
+{-@ predicate ValidDirectorScheme V = (listElts (ddom V) = (Set_sng Director)) @-} 
+{-@ predicate DirectorDomain T   = ( T = Director ) @-}
+{-@ predicate DirectorRange  T V = (T = Director => ValidDirector V) @-}
 
-{-@ predicate ValidStarScheme V = (listElts (ddom V) = (Set_sng star)) @-} 
-{-@ predicate StarDomain T      = (T = star) @-}
-{-@ predicate StarRange  T V    = (T = star => ValidStar V) @-}
+{-@ predicate ValidStarScheme V = (listElts (ddom V) = (Set_sng Star)) @-} 
+{-@ predicate StarDomain T      = (T = Star) @-}
+{-@ predicate StarRange  T V    = (T = Star => ValidStar V) @-}
 
-{-@ predicate ValidTitleScheme V = (listElts (ddom V) = (Set_sng title)) @-} 
-{-@ predicate TitleDomain T   = ( T = title) @-}
-{-@ predicate TitleRange  T V = (T = title => ValidTitle V) @-}
+{-@ predicate ValidTitleScheme V = (listElts (ddom V) = (Set_sng Title)) @-} 
+{-@ predicate TitleDomain T   = ( T = Title) @-}
+{-@ predicate TitleRange  T V = (T = Title => ValidTitle V) @-}
 
-{-@ predicate ValidDirStarScheme V = (listElts (ddom V) = Set_cup (Set_sng director) (Set_sng star)) @-} 
-{-@ predicate DirStarDomain T   = ( T = director || T = star) @-}
-{-@ predicate DirStarRange  T V = (T = director => ValidDirector V) && (T = star => ValidStar V)  @-}
+{-@ predicate ValidDirStarScheme V = (listElts (ddom V) = Set_cup (Set_sng Director) (Set_sng Star)) @-} 
+{-@ predicate DirStarDomain T   = ( T = Director || T = Star) @-}
+{-@ predicate DirStarRange  T V = (T = Director => ValidDirector V) && (T = Star => ValidStar V)  @-}
 
 
 {-@ predicate ValidYear     V = isInt V  && 1889 <= toInt V  @-}
@@ -96,42 +94,42 @@ mkMovie :: Value -> Value -> Value -> Value -> MovieScheme
             -> {y:Value | ValidYear y}
             -> MovieScheme
   @-}          
-mkMovie t d s y = ("title" := t) += ("star" := s) += ("director" := d) += ("year" := y) += empty
+mkMovie t d s y = (Title := t) += (Star := s) += (Director := d) += (Year := y) += empty
 
 seen :: Titles
 {-@ seen :: Titles @-}
 seen = [t1, t2]
   where 
-  	t1 = ("title" := S ChickenPlums) += empty
-  	t2 = ("title" := S FunnyGames)   += empty
+  	t1 = (Title := S ChickenPlums) += empty
+  	t2 = (Title := S FunnyGames)   += empty
 
 not_seen :: Movies
 not_seen = select isSeen movies
   where
-   	isSeen t v | t == "title" = not $ v `elem` (values "title" seen)
-  	isSeen _ _                = True
+  	isSeen Title t = not $ t `elem` (values Title seen)
+  	isSeen _     _ = True
 
 to_see = select isGoodMovie not_seen
   where
-   isGoodMovie t (I s) | t == "star"     = s >= 8
-   isGoodMovie t d     | t == "director" = d `elem` (values "director" good_directors)
-   isGoodMovie _ _                       = True
+   isGoodMovie Star     (I s) = s >= 8
+   isGoodMovie Director d     = d `elem` (values Director good_directors)
+   isGoodMovie _        _     = True
 
 directors, good_directors :: Directors
 {-@ directors, good_directors :: Directors @-}
-directors = project ["director"] movies
+directors = project [Director] movies
 
 
 good_stars :: Stars
 {-@ good_stars :: Stars @-}
-good_directors     = directors `diff` project ["director"] not_good_directors
+good_directors     = directors `diff` project [Director] not_good_directors
 -- This _IS_ unsafe!
 -- good_directors     = directors `diff` not_good_directors
 
 not_good_directors :: DirStars 
 {-@ not_good_directors :: DirStars @-}
-not_good_directors = project ["director", "star"] movies  `diff` product directors good_stars 
-
+-- not_good_directors = project [Director, Star] movies  `diff` [ productD x y | x <- directors, y <- good_stars] 
+not_good_directors = project [Director, Star] movies  `diff` product directors good_stars -- [ productD x y | x <- directors, y <- good_stars] 
  
 -- This _IS_ unsafe! 
 -- not_good_directors = project [Director, Star] movies  `diff` [ productD x y | x <- directors, y <- movies] 
@@ -140,7 +138,7 @@ good_stars         = mk_star_table (I 8) `union` mk_star_table (I 9) `union` mk_
 
 mk_star_table :: Value -> Stars
 {-@ mk_star_table :: {s:Value | ValidStar s} -> Stars @-}
-mk_star_table s    = [("star" := s) += empty]
+mk_star_table s    = [(Star := s) += empty]
 
 
 -------------------------------------------------------------------------------
